requirements.txt
attrs==18.2.0
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
distro==1.4.0
idna==2.8
lxml==4.3.1
more-itertools==6.0.0
numpy==1.16.1
pandas==0.24.1
pluggy==0.9.0
ply==3.11
py==1.8.0
pytest==4.3.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.8
tabula-py==1.3.1
urllib3==1.24.1
